SHADER: changed position to be 3-component byte vector

+++ data/shared/shaders/world.vert
@@ -1,5 +1,5 @@
 // attributes from the VAOs
-in vec3 a_pos;
+in uvec3 a_pos;
 in int a_ao;
 in int a_material;
